// frontend/src/api/generated/model/updateStartupDtoRaiseType.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export type UpdateStartupDtoRaiseType = typeof UpdateStartupDtoRaiseType[keyof typeof UpdateStartupDtoRaiseType];


export const UpdateStartupDtoRaiseType = {
  safe: 'safe',
  convertible_note: 'convertible_note',
  equity: 'equity',
  safe_equity: 'safe_equity',
  undecided: 'undecided',
} as const;