// frontend/src/api/generated/model/updateStartupDtoStage.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * NestJS Boilerplate API
 * Production-ready NestJS boilerplate with auth, queues, storage, and email
 * OpenAPI spec version: 2.0
 */

export type UpdateStartupDtoStage = typeof UpdateStartupDtoStage[keyof typeof UpdateStartupDtoStage];


export const UpdateStartupDtoStage = {
  pre_seed: 'pre_seed',
  seed: 'seed',
  series_a: 'series_a',
  series_b: 'series_b',
  series_c: 'series_c',
  series_d: 'series_d',
  series_e: 'series_e',
  series_f_plus: 'series_f_plus',
} as const;