// frontend/src/api/generated/model/updateStartupDtoValuationType.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Inside Line API
 * Inside Line API Documentation
 * OpenAPI spec version: 2.0
 */

export type UpdateStartupDtoValuationType = typeof UpdateStartupDtoValuationType[keyof typeof UpdateStartupDtoValuationType];


export const UpdateStartupDtoValuationType = {
  pre_money: 'pre_money',
  post_money: 'post_money',
} as const;